pyproject.toml
[tool.poetry]
name = "github-notifications-cleaner"
version = "0.0.2"
description = ""
authors = [
    "James Hogkinson <james@terminaloutcomes.com>",
]
license = "MIT"

[tool.poetry.scripts]
github-notifications-cleaner = "github_notifications_cleaner.__main__:cli"

[tool.poetry.dependencies]
python = "^3.9"
PyGithub = "^1.55"
click = "^8.1.3"
loguru = "^0.6.0"

[tool.poetry.dependencies.pydantic]
extras = [
    "dotenv",
]
version = "^1.10.1"

[tool.poetry.dev-dependencies]
pylint = "^2.15.2"
mypy = "^0.971"
pytest = "^7.1.3"
black = "^22.8.0"

[tool.pylint.MASTER]
max-line-length = 200

[build-system]
requires = [
    "poetry-core>=1.0.0",
]
build-backend = "poetry.core.masonry.api"
